requirements.txt
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
SpeechRecognition==3.10.0
PyAudio==0.2.13
